Move models and losses to root of folder
